% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genocut.R
\name{genocut}
\alias{genocut}
\title{Genocut}
\usage{
genocut(file, len = 0, num = 0)
}
\arguments{
\item{file}{FASTA file}

\item{len}{length of the sequence to be cut}

\item{num}{Number of sequence to be generated}
}
\value{
character vector
}
\description{
Genocut
}
\examples{
dat <- genocut ("test.fas", 300, 10)

}
